+++ include/Binary/BinaryDeserializer.hpp
@@ -106,12 +106,12 @@ namespace claid
                 member = static_cast<T>(m);
             }
 
-            void count(const std::string& name, size_t& count)
+            void count(const std::string& name, int32_t& count)
             {
                 this->binaryDataReader.read(count);
             }
 
-            void countElements(size_t& count)
+            void countElements(int32_t& count)
             {
                 this->binaryDataReader.read(count);
             }
